Compile also on ESP32 based boards
Also compiles on ESP32 (S2/C3), so Adafruit QT PY, Seeed Studio XIAO, and other alternatives to the Digistump.

The avr/pgmspace.h does not need to be included anymore, it's done automatically if a AVR core is chosen. And for the ES32 boards PROGMEM just gets ignored, but they have more than enough memory.